requirements.txt
fastapi>=0.100.0
mangum>=0.17.0
boto3>=1.28.0
google-api-python-client>=2.90.0
google-auth>=2.22.0
dropbox>=11.36.0
pydantic>=2.0.0
pydantic-settings>=2.0.0
uvicorn>=0.23.0
httpx>=0.24.0
